requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
PyMuPDF==1.24.11
python-docx==1.1.2
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.4.0
groq==0.9.0
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
